requirements.txt
Flask==2.1.2
torch==2.0.1
Pillow==9.2.0
numpy==1.23.1
opencv-python==4.6.0.66
diffusers==0.8.0
transformers==4.21.0
accelerate==0.12.0
scipy==1.8.0
safetensors==0.2.0
huggingface-hub==0.8.1
torchvision==0.14.0
Werkzeug==2.1.2
